+++ src/main/php/lang/mirrors/Sources.class.php
@@ -3,9 +3,9 @@
 /**
  * Sources from which reflection can be created:
  *
- * - DEFAULT: Uses reflection if class exists, parsing code otherwise
- * - REFLECTION: Uses ext/reflection
- * - CODE: Parses code.
+ * - DEFAULT: Same as REFLECTION
+ * - REFLECTION: Uses reflection if class exists, parsing code otherwise
+ * - CODE: Parses code if available, using reflection otherwise
  *
  * Has special case handling to cope with situation that class is not
  * fully defined (e.g. when performing compile-time metaprogramming).
@@ -16,37 +16,26 @@ abstract class Sources extends \lang\Enum {
 
   static function __static() {
     $reflect= defined('HHVM_VERSION') ? 'FromHHVM' : 'From';
-    self::$DEFAULT= newinstance(self::class, [0, 'DEFAULT'], sprintf('{
-      static function __static() { }
-
-      public function reflect($class, $source= null) {
-        if ($class instanceof \ReflectionClass) {
-          return new %1$sReflection($class, $source ?: $this);
-        }
-
-        $literal= strtr($class, ".", "\\\\");
-        $dotted= strtr($class, "\\\\", ".");
-        if (class_exists($literal) || interface_exists($literal) || trait_exists($literal)) {
-          return self::$REFLECTION->reflect($class, $source ?: $this);
-        } else if (\lang\ClassLoader::getDefault()->providesClass($dotted)) {
-          return new %1$sCode($dotted, $source ?: $this);
-        } else {
-          return new FromIncomplete($literal);
-        }
-      }
-    }', $reflect));
     self::$REFLECTION= newinstance(self::class, [1, 'REFLECTION'], sprintf('{
       static function __static() { }
 
       public function reflect($class, $source= null) {
         if ($class instanceof \ReflectionClass) {
-          return new %1$sReflection($class, $source);
+          return new %1$sReflection($class, $source ?: $this);
+        } else if ($class instanceof \lang\XPClass) {
+          return new %1$sReflection($class->reflect(), $source ?: $this);
         }
 
-        try {
-          return new %1$sReflection(new \ReflectionClass(strtr($class, ".", "\\\\")), $source);
-        } catch (\Exception $e) {
-          throw new \lang\ClassNotFoundException($class.": ".$e->getMessage());
+        $literal= strtr($class, ".", "\\\\");
+        if (class_exists($literal) || interface_exists($literal) || trait_exists($literal)) {
+          return new %1$sReflection(new \ReflectionClass($literal), $source ?: $this);
+        }
+
+        $dotted= strtr($class, "\\\\", ".");
+        if (\lang\ClassLoader::getDefault()->providesClass($dotted)) {
+          return new %1$sCode($dotted, $source ?: $this);
+        } else {
+          return new FromIncomplete($literal);
         }
       }
     }', $reflect));
@@ -54,18 +43,35 @@ abstract class Sources extends \lang\Enum {
       static function __static() { }
 
       public function reflect($class, $source= null) {
-        return new %1$sCode(strtr($class, "\\\\", "."), $source);
+        if ($class instanceof \ReflectionClass) {
+          return new %1$sReflection($class, $source ?: $this);
+        } else if ($class instanceof \lang\XPClass) {
+          return new %1$sReflection($class->reflect(), $source ?: $this);
+        }
+
+        $dotted= strtr($class, "\\\\", ".");
+        if (\lang\ClassLoader::getDefault()->providesClass($dotted)) {
+          return new %1$sCode($dotted, $source ?: $this);
+        }
+
+        $literal= strtr($class, ".", "\\\\");
+        if (class_exists($literal) || interface_exists($literal) || trait_exists($literal)) {
+          return new %1$sReflection(new \ReflectionClass($literal), $source ?: $this);
+        } else {
+          return new FromIncomplete($literal);
+        }
       }
     }', $reflect));
+
+    self::$DEFAULT= self::$REFLECTION;
   }
 
   /**
    * Creates a reflection source for a given class
    *
-   * @param  string $class
+   * @param  var $class Either a lang.XPClass, a ReflectionClass or a class name
    * @param  self $source
    * @return lang.mirrors.Source
-   * @throws lang.ClassNotFoundException
    */
   public abstract function reflect($class, $source= null);
 }